% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{max_stable_prcomp}
\alias{max_stable_prcomp}
\title{Calculate max-stable PCA with dimension p for given dataset}
\usage{
max_stable_prcomp(data, p, s = 3, x0 = NULL)
}
\arguments{
\item{data, }{array or data.frame of n observations of d variables
with unit Frechet margins. The max-stable PCA is fitted to
reconstruct this dataset with a rank p approximation.}

\item{p, }{integer between 1 and ncol(data). Determines
the dimension of the encoded state, i.e. the number of max-linear
combinations in the compressed representation.}

\item{s}{(default = 3), numeric greater than 0. Hyperparameter for the
stable tail dependence estimator used in tn the calculation.}

\item{x0}{(default = NULL), initial value for optimizer. Set with
extreme caution and only with good knowledge of the minimizationb problem.
NOTE: will be moved to ...}
}
\value{
object of class max_stable_prcomp with slots
TODO
}
\description{
Find
}
\examples{
# TODO
1 + 1
}
