% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{reconstr_loss}
\alias{reconstr_loss}
\title{Loss for a given reconstruction matrix}
\usage{
reconstr_loss(H, data, s)
}
\arguments{
\item{H, }{a (d x d) matrix where d is the number of columns of data}

\item{data, }{a matrix like object of shape (n x d), where d is the same as for H. Should be transformed (TODO).}

\item{s, }{non-negative numeric, tunig parameter for the stable tail dependence function. For further information see TODO.}
}
\description{
TODO
}
\examples{
TODO
}
