% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct the data from the compressed representation}
\usage{
reconstruct(fit, compressed_data)
}
\arguments{
\item{fit, }{max_stable_prcomp object. Data should be
assumed to follow the same distribution as the data used in
max_stable_prcomp.}
}
\value{
array of dimension nrow(compressed_data), d
An optimal reconstruction with max-linear combinations 
from p max-linearly independent columns.
}
\description{
Calculate the reconstruction given by the compressed state
compressed_data and the fit object from the max-stable PCA.
Can be used for visual assessment of the fit if p should be chosen
differently.
}
\examples{
# TODO
1 + 1
}
\seealso{
[maxstablePCA::compress(), 
maxstablePCA::max_stable_prcomp(),
maxstablePCA::maxmatmul()]
}
