% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_taildep.R
\name{stable_tail_dependence}
\alias{stable_tail_dependence}
\title{Calculate the estimated stable tail dependence at point x for given dataset.}
\usage{
stable_tail_dependence(x, s, data)
}
\arguments{
\item{x, }{array with ncol(data) entries that is non-negative.}

\item{s, }{double that is non-negative. Controls the cutoff of data for estimator}

\item{data, }{array where the rows are representing the observations and 
the columns are the points sampled from. The columns are assumed to 
have \emph{unit Frechet} marginals.}
}
\value{
Non negative float, representing the stable tail depndence 
estimator evaluated at parameter x.
}
\description{
The stable tail dependence is a function describing the 
full dependence structure of a multivariate max-stable random
vector with unit Frechet margins (loc = 0, scale = 1, shape = 1). 
It takes a similar place to the covariance for the gaussian distribution.
}
\details{
The exact formula and its theoretical properties can be found in
Beirlant, Jan. Statistics of Extremes. (2004). Formula (9.60) is
the estimator.
}
\examples{
# set up example data
data <- matrix(c(
 69.9399283, 3.960401, 69.939928,
 0.3952988, 4.416034,  4.416034,
 4.9373416,  4.937342,  1.698470,
 0.8265868,  8.180792,  8.180792
), 3, 3)

# call function for 3 variate random vector sample
stable_tail_dependence(c(1,2,3), 3, data)
}
\seealso{
[maxstablePCA::transform_unitfrechet()] to make sure the marginal 
distributions of the columns are standardized.
}
