% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_taildep.R
\name{stable_tail_dependence_univC}
\alias{stable_tail_dependence_univC}
\title{Intern function for stable tail dependence before vectorization}
\usage{
stable_tail_dependence_univC(x, s, data)
}
\arguments{
\item{x, }{array with ncol(data) entries that is non-negative.}

\item{s, }{double that is non-negative. Controls the cutoff of data for estimator}

\item{data, }{array where the rows are representing the observations and 
the columns are the points sampled from. The columns are assumed to 
have \emph{unit Frechet} marginals.}
}
\value{
Non negative float, representing the stable tail depndence 
estimator evaluated at parameter x.
}
\description{
Intern function for stable tail dependence before vectorization
}
