% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{summary.max_stable_prcomp}
\alias{summary.max_stable_prcomp}
\title{Print summary of a max_stable_prcomp object.}
\usage{
\method{summary}{max_stable_prcomp}(object, ...)
}
\arguments{
\item{object, }{max_stable_prcomp object. Data should be
assumed to follow the same distribution as the data used in
max_stable_prcomp.}

\item{...}{additional unused arguments.}
}
\description{
Print summary of a max_stable_prcomp object.
}
\examples{
# Alternatively call via the generic function summary
# TODO: generic fit example
# summary.max_stable_prcomp(fit)
}
