% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_unitfrechet.R
\name{transform_orig_margins}
\alias{transform_orig_margins}
\title{Transform the columns of a transformed dataset to original margins}
\usage{
transform_orig_margins(transformed_data, orig_data, empirical = TRUE)
}
\arguments{
\item{A}{TODO}

\item{B}{TODO}
}
\value{
TODO
}
\description{
Lore ipsum
}
\examples{
# TODO
1 + 1
}
\seealso{
[maxstablePCA::max_stable_prcomp()] for information about why to transform data.
}
