% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_unitfrechet.R
\name{transform_unitfrechet}
\alias{transform_unitfrechet}
\title{Transform the columns of a dataset to unit Frechet margins}
\usage{
transform_unitfrechet(data, empirical = TRUE)
}
\arguments{
\item{data, }{array or vector with the data which columns are to be transformed}

\item{empirical, }{logical to indicate the transformation method. If true the data
is transformed using the empical distribution function. If \code{FALSE}, the 
data is transformed using a MLE fit of a generalized extreme value distribution to the data.
Using the empirical distribution function does not require the data to have approximately
max-stable distribution, thus is set to default.}
}
\value{
array or vector of same shape and type as data with the transformed data with unit 
Freceht margins-
}
\description{
Lore ipsum
}
\examples{
# TODO
1 + 1
}
\seealso{
[maxstablePCA::max_stable_prcomp()] for information about why to transform data.
}
