% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_unitfrechet.R
\name{transform_unitpareto}
\alias{transform_unitpareto}
\title{Transform the columns of a dataset to unit Pareto}
\usage{
transform_unitpareto(data, empirical = TRUE, ...)
}
\arguments{
\item{data, }{array or vector with the data which columns are to be transformed}

\item{empirical, }{logical to indicate the transformation method. If true the data
is transformed using the empical distribution function. If \code{FALSE}, the 
data is transformed using a MLE fit of a generalized extreme value distribution to the data.
Using the empirical distribution function does not require the data to have approximately
max-stable distribution, thus is set to default.}

\item{..., }{additional arguments for \code{mev::fit.gev()}}
}
\value{
array or vector of same shape and type as data with the transformed data with unit 
Frechet margins-
}
\description{
Transforms columns of dataset to unit Pareto margins, to ensure
the theoretical requirements are satisfied for the application of
\code{\link{max_stable_prcomp}}.
}
\details{
Additional care is needed when using the non empirical method, since for large 
datasets the optimizer may not converge for some columns and the assumption of a GEV
on the marginals should be carefully checked.
}
\examples{
# sample some data
dat <- rnorm(1000)
transformed_dat <- transform_unitfrechet(dat)

# Look at a plot of distribution
boxplot(transformed_dat)
plot(stats::ecdf(transformed_dat))
}
\seealso{
[maxstablePCA::max_stable_prcomp(), maxstablePCA::transform_orig_margins(), mev::fit.gev())] for information about why to transform data
and the MLE method used internally.
}
